View, ViewModel, Model 분리
